greet: Added client and server Greet implementation using Unary gRPC
Server:
1. GreetServiceServer is the server API for GreetService service
The above is an interface type. Hence we need to implement this
interface in our server object. Which I guess will then be called by
gRPC automatically since it is attached to server struct which was
passed while greetpb.RegisterGreetServiceServer(s, &server{})

Client:
1. GreetServiceClient is the client API for GreetService service.
We now need to call c.Greet() and pass the req to the server.
We in turn should get a response back from the server for this.

Signed-off-by: Ritesh Harjani <ritesh.harjani@gmail.com>